+++ examples/features/src/stencil_triangles/mod.rs
@@ -45,8 +45,11 @@ impl crate::framework::Example for Example {
             usage: wgpu::BufferUsages::VERTEX,
         });
 
-        let pipeline_layout =
-            device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor::empty());
+        let pipeline_layout = device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
+            label: None,
+            bind_group_layouts: &[],
+            immediate_size: 0,
+        });
 
         let shader = device.create_shader_module(wgpu::include_wgsl!("shader.wgsl"));