+++ tests/tests/wgpu-gpu/binding_array/buffers.rs
@@ -227,7 +227,7 @@ async fn binding_array_buffers(
         .device
         .create_pipeline_layout(&PipelineLayoutDescriptor {
             label: Some("Pipeline Layout"),
-            bind_group_layouts: &[&bind_group_layout],
+            bind_group_layouts: &[Some(&bind_group_layout)],
             immediate_size: 0,
         });