+++ tests/tests/wgpu-gpu/binding_array/samplers.rs
@@ -213,7 +213,7 @@ async fn binding_array_samplers(ctx: TestingContext, partially_bound: bool) {
         .device
         .create_pipeline_layout(&PipelineLayoutDescriptor {
             label: Some("Pipeline Layout"),
-            bind_group_layouts: &[&bind_group_layout],
+            bind_group_layouts: &[Some(&bind_group_layout)],
             immediate_size: 0,
         });