+++ tests/tests/wgpu-gpu/mesh_shader/mod.rs
@@ -210,7 +210,11 @@ fn mesh_pipeline_build(ctx: &TestingContext, info: MeshPipelineTestInfo) {
         ms_name,
         fs_name,
     } = get_shaders(device, backend, &test_hash, &info);
-    let layout = device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor::empty());
+    let layout = device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
+        label: None,
+        bind_group_layouts: &[],
+        immediate_size: 0,
+    });
     let pipeline = device.create_mesh_pipeline(&wgpu::MeshPipelineDescriptor {
         label: None,
         layout: Some(&layout),
@@ -292,7 +296,11 @@ fn mesh_draw(ctx: &TestingContext, draw_type: DrawType, info: MeshPipelineTestIn
         fs_name,
     } = get_shaders(device, backend, &test_hash, &info);
     let frag = fs.unwrap();
-    let layout = device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor::empty());
+    let layout = device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
+        label: None,
+        bind_group_layouts: &[],
+        immediate_size: 0,
+    });
     let pipeline = device.create_mesh_pipeline(&wgpu::MeshPipelineDescriptor {
         label: None,
         layout: Some(&layout),