+++ tests/tests/wgpu-gpu/regression/issue_3349.rs
@@ -92,7 +92,7 @@ async fn multi_stage_data_binding_test(ctx: TestingContext) {
         .device
         .create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
             label: Some("pll"),
-            bind_group_layouts: &[&bgl],
+            bind_group_layouts: &[Some(&bgl)],
             immediate_size: 16,
         });