+++ CHANGELOG.md
@@ -44,6 +44,25 @@ Bottom level categories:
 
 ### Major Changes
 
+#### Bind group layouts now optional in `PipelineLayoutDescriptor`
+
+Allow gaps in bind group layouts and added full support for unbinding. As a result of this `PipelineLayoutDescriptor`'s `bind_group_layouts` field now has type of `&[Option<&BindGroupLayout>]`, making this a breaking change. To migrate wrap bind group layout references in `Some`:
+
+```diff
+  let pl_desc = wgpu::PipelineLayoutDescriptor {
+      label: None,
+      bind_group_layouts: &[
+-         &bind_group_layout
++         Some(&bind_group_layout)
+      ],
+      immediate_size: 0,
+  });
+```
+
+By @teoxoy in [#9034](https://github.com/gfx-rs/wgpu/pull/9034).
+
+#### MSRV update
+
 `wgpu` now has a new MSRV policy. This release has an MSRV of **1.87**. This is lower than v27's 1.88 and v28's 1.92. Going forward, we will only bump wgpu's MSRV if it has tangible benefits for the code, and we will never bump to an MSRV higher than `stable - 3`. So if stable is at 1.97 and 1.94 brought benefit to our code, we could bump it no higher than 1.94. As before, MSRV bumps will always be breaking changes. 
 
 By @cwfitzgerald in [#8999](https://github.com/gfx-rs/wgpu/pull/8999).
@@ -103,7 +122,6 @@ depth_stencil: Some(wgpu::DepthStencilState::stencil(
 - Added support for `enable primitive_index` and `@builtin(primitive_index)` with support on all platforms. By @inner-daemons in #8879.
 - BREAKING: Add `V6_8` variant to `DxcShaderModel` and `naga::back::hlsl::ShaderModel`. By @inner-daemons in [#8882](https://github.com/gfx-rs/wgpu/pull/8882) and @ErichDonGubler in [#9083](https://github.com/gfx-rs/wgpu/pull/9083).
 - BREAKING: Add `V6_9` variant to `DxcShaderModel` and `naga::back::hlsl::ShaderModel`. By @ErichDonGubler in [#????](https://github.com/gfx-rs/wgpu/pull/????).
-- Allow gaps in bind group layouts and added full support for unbinding. By @teoxoy in [#9034](https://github.com/gfx-rs/wgpu/pull/9034).
 
 #### naga